requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.7
pandas==2.2.3
python-dotenv==1.0.0